requirements.txt
flask
flask-cors
openai
python-dotenv
feedparser
requests
gunicorn
httpx
beautifulsoup4